+++ internal/interpreter/zz_interpreter_get_test.go
@@ -79,7 +79,7 @@ func TestGET_TableDriven(t *testing.T) {
 		},
 	}
 
-	rt, _ := machines.NewRuntime(constants.BASIC_TTY)
+	rt, _ := machines.NewRuntime(constants.BASIC_TTY, false)
 
 	for tIndex, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {