+++ internal/interpreter/zz_interpreter_input_test.go
@@ -95,7 +95,7 @@ func TestINPUT_TableDriven(t *testing.T) {
 		}, */
 	}
 
-	rt, _ := machines.NewRuntime(constants.BASIC_TTY)
+	rt, _ := machines.NewRuntime(constants.BASIC_TTY, false)
 
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {