+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog (https://keepachangelog.com), and this project adheres to Semantic Versioning (https://semver.org).
 
+## [Unreleased] - 2026-02-19
+### Added
+- Add `PR#3` support in Apple II Basic to switch to 80 columns mode.
+
 ## [Unreleased] - 2026-02-18
 ### Fixed
 - Fix a bug when testing strings for equality.